fix: update id

+++ src/main/java/maxim/butenko/weather/dto/UserDTO.java
@@ -1,13 +1,16 @@
 package maxim.butenko.weather.dto;
 
 import lombok.Builder;
+import lombok.NoArgsConstructor;
 import lombok.Value;
 import maxim.butenko.weather.entity.Role;
+
+import java.math.BigInteger;
 
 @Value
 @Builder
 public class UserDTO {
-    Long id;
+    BigInteger id;
     String login;
     String password;
     Role role;
